isl/requirements.txt


Django
djangorestframework
psycopg2-binary
graphene-django